python-app/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
flake8==4.0.1
frozenlist==1.3.0
idna==3.3
kafka-python==2.0.2
mccabe==0.6.1
multidict==6.0.2
pycodestyle==2.8.0
pyflakes==2.4.0
python-decouple==3.6
requests==2.27.1
toml==0.10.2
urllib3==1.26.9
yarl==1.7.2
